+++ tests/test_general.py
@@ -17,7 +17,7 @@ import os
 
 
 def test_version():
-    assert "1.26.0" == deepl.__version__
+    assert "1.27.0" == deepl.__version__
 
 
 @pytest.mark.parametrize(