+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.27.0] - 2025-12-10
 ### Added
 - Added `tag_handling_version` parameter to `translate_text()` to specify which version
   of the tag handling algorithm to use. Options are `v1` and `v2`.
@@ -362,7 +364,8 @@ Version increased to avoid conflicts with old packages on PyPI.
 ## [0.1.0] - 2021-07-26
 Initial version.
 
-[Unreleased]: https://github.com/DeepLcom/deepl-python/compare/v1.26.0...HEAD
+[Unreleased]: https://github.com/DeepLcom/deepl-python/compare/v1.27.0...HEAD
+[1.27.0]: https://github.com/DeepLcom/deepl-python/compare/v1.26.0...v1.27.0
 [1.26.0]: https://github.com/DeepLcom/deepl-python/compare/v1.25.0...v1.26.0
 [1.25.0]: https://github.com/DeepLcom/deepl-python/compare/v1.24.0...v1.25.0
 [1.24.0]: https://github.com/DeepLcom/deepl-python/compare/v1.23.0...v1.24.0